requirements.txt
py-cord
Flask
langdetect
python-dotenv
googletrans==4.0.0rc1
replicate
aiohttp
requests
ollama==0.4.7